keep reasonable default values for app:entries-xpath... variables
We can keep the defaults for app:entries-xpath-... and leave the
overrideable at the same time. There's no reason to make them abstract
because they do not depend on a global context item. So I've switched
their visibility to 'public'.

This makes the package executable, which is nice for testing.

And it make the package a super-easy plugin.